refactored, parity with 6.5.0
