+++ OneMore/OneMore.csproj
@@ -226,7 +226,10 @@
     <Compile Include="Commands\Settings\ImagesSheet.Designer.cs">
       <DependentUpon>ImagesSheet.cs</DependentUpon>
     </Compile>
+    <Compile Include="Commands\Snippets\TocGenerators\HierarchyTocGenerator.cs" />
+    <Compile Include="Commands\Snippets\TocGenerators\NotebookTocGenerator.cs" />
     <Compile Include="Commands\Snippets\TocGenerators\PageTocGenerator.cs" />
+    <Compile Include="Commands\Snippets\TocGenerators\SectionTocGenerator.cs" />
     <Compile Include="Commands\Snippets\TocGenerators\TocCommon.cs" />
     <Compile Include="Commands\Snippets\TocGenerators\TocGenerator.cs" />
     <Compile Include="Commands\Snippets\TocGenerators\TocParameters.cs" />
